node/mempool: PR #1422 wave-9 split precheck helpers to mempool_precheck.go (Codacy file-cumulative complexity budget)
Wave-9 closes the remaining Codacy "Complexity increased by 25
(20 max.)" file-level warning by moving the precheck/admit-validation
block from mempool.go to a new sibling file mempool_precheck.go in
the same package. Same package = no API change, no test impact.

Functions moved (lines 667-955 of mempool.go pre-split):
- buildPolicyInputSnapshotIfNeeded (wave-8)
- (m *Mempool) checkTxParseAndContext (wave-7)
- (m *Mempool) checkTransactionWithSnapshot (the orchestration that
  uses the helpers)
- cheapFeeFloorPrecheck (wave-1..6)
- precheckEarlyDefer (wave-7)
- feePrecheckP2PKInputValue (wave-1..5)
- precheckP2PKInputStructurallyValid (wave-7)
- precheckCoinbaseImmature (wave-5/7)
- feePrecheckP2PKOutputValue (wave-1..4)

mempool.go now sheds ~290 lines and the file-cumulative complexity
contribution from those helpers, while preserving every per-method
complexity budget (already met after wave-7/8 extracts). All other
parts of the admission pipeline (addTxWithSource, addEntryLocked,
validateFeeFloorLocked, addEntryLockedWithFloor,
validateFeeFloorLockedWithFloor, etc.) stay in mempool.go.

Wave-9 cumulative gate verification:
- gofmt -l ./node clean
- go vet ./node/... clean
- go build ./node/... clean
- go test ./node -run TestMempool -count=1 PASS
